requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0
trl>=0.8.0
einops>=0.7.0

# Medical Imaging
Pillow>=10.0.0
pydicom>=2.4.0
opencv-python>=4.8.0
open-clip-torch>=2.24.0

# Data & Science
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Experiment Tracking
mlflow>=2.11.0
wandb>=0.16.0

# Config Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Evaluation & Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.0

# Data Versioning
dvc>=3.40.0

# Serving
fastapi>=0.110.0
uvicorn>=0.27.0

# Dev
pytest>=8.0.0
black>=24.0.0
ruff>=0.3.0
pre-commit>=3.6.0
ipykernel>=6.29.0
python-dotenv>=1.0.0
